fix bugs, add services and controller

+++ backend/src/main/kotlin/com/bftcom/backend/entity/LibraryBook.kt
@@ -2,5 +2,5 @@ package com.bftcom.backend.entity
 
 data class LibraryBook(
 	override var id: Long = 0,
-	val bookId: Int
+	val bookId: Long
 ) : Entity
